Replace Data with Setup.data
The `countries` gem has updated the `ISO3166::Country class` in version
0.11.5 so that you need to call `Setup.data` instead of `Data`.

In order to be able to run this gem locally, I had to upgrade coveralls
because term-ansicolor 1.3.1 no longer exists.

I also updated the `countries` gem dependency in this gemspec since this
new code will only work with 0.11.5 and above.

Closes #103.